Replace RadialGradient with ConicalGradient
